refactor supCon.py
move loss function in supCon.py to utils/losses.py
