+++ sample_supCon.py
@@ -15,6 +15,7 @@ from utils.common import time_counter
 
 from models.simple import SimpleCNN
 from models.wrapper import VGG16, ResNet50
+from utils.losses import SupConLoss
 
 tfk = tf.keras
 tfk.backend.set_floatx('float32')
@@ -103,132 +104,14 @@ optimizer = tfk.optimizers.Adam(learning_rate=1e-3)
 encoder_r = encoder_net()
 projector_z = projector_net()
 
-# =============================================================================
-# define max_margin_contrastive_loss
-# 以下のファイルからコピペ
-# tensorflow_addons, https://raw.githubusercontent.com/wangz10/contrastive_loss/master/losses.py
-# =============================================================================
-def pdist_euclidean(A):
-    # Euclidean pdist
-    # https://stackoverflow.com/questions/37009647/compute-pairwise-distance-in-a-batch-without-replicating-tensor-in-tensorflow
-    r = tf.reduce_sum(A*A, 1)
-
-    # turn r into column vector
-    r = tf.reshape(r, [-1, 1])
-    D = r - 2*tf.matmul(A, tf.transpose(A)) + tf.transpose(r)
-    return tf.sqrt(D)
-
-def square_to_vec(D):
-    '''Convert a squared form pdist matrix to vector form.
-    '''
-    n = D.shape[0]
-    triu_idx = np.triu_indices(n, k=1)
-    d_vec = tf.gather_nd(D, list(zip(triu_idx[0], triu_idx[1])))
-    return d_vec
-
-def get_contrast_batch_labels(y):
-    '''
-    Make contrast labels by taking all the pairwise in y
-    y: tensor with shape: (batch_size, )
-    returns:   
-        tensor with shape: (batch_size * (batch_size-1) // 2, )
-    '''
-    y_col_vec = tf.reshape(tf.cast(y, tf.float32), [-1, 1])
-    D_y = pdist_euclidean(y_col_vec)
-    d_y = square_to_vec(D_y)
-    y_contrasts = tf.cast(d_y == 0, tf.int32)
-    return y_contrasts
-
-from typing import Union, List
-
-import numpy as np
-import tensorflow as tf
-
-Number = Union[
-    float,
-    int,
-    np.float16,
-    np.float32,
-    np.float64,
-    np.int8,
-    np.int16,
-    np.int32,
-    np.int64,
-    np.uint8,
-    np.uint16,
-    np.uint32,
-    np.uint64,
-]
-
-TensorLike = Union[
-    List[Union[Number, list]],
-    tuple,
-    Number,
-    np.ndarray,
-    tf.Tensor,
-    tf.SparseTensor,
-    tf.Variable,
-]
-@tf.function
-def contrastive_loss(
-    y_true: TensorLike, y_pred: TensorLike, margin: Number = 1.0
-) -> tf.Tensor:
-    r"""Computes the contrastive loss between `y_true` and `y_pred`.
-    This loss encourages the embedding to be close to each other for
-    the samples of the same label and the embedding to be far apart at least
-    by the margin constant for the samples of different labels.
-    The euclidean distances `y_pred` between two embedding matrices
-    `a` and `b` with shape [batch_size, hidden_size] can be computed
-    as follows:
-    ```python
-    # y_pred = \sqrt (\sum_i (a[:, i] - b[:, i])^2)
-    y_pred = tf.linalg.norm(a - b, axis=1)
-    ```
-    See: http://yann.lecun.com/exdb/publis/pdf/hadsell-chopra-lecun-06.pdf
-    Args:
-      y_true: 1-D integer `Tensor` with shape [batch_size] of
-        binary labels indicating positive vs negative pair.
-      y_pred: 1-D float `Tensor` with shape [batch_size] of
-        distances between two embedding matrices.
-      margin: margin term in the loss definition.
-    Returns:
-      contrastive_loss: 1-D float `Tensor` with shape [batch_size].
-    """
-    y_pred = tf.convert_to_tensor(y_pred)
-    y_true = tf.dtypes.cast(y_true, y_pred.dtype)
-    return y_true * tf.math.square(y_pred) + (1.0 - y_true) * tf.math.square(
-        tf.math.maximum(margin - y_pred, 0.0)
-    )
-
-def max_margin_contrastive_loss(z, y, margin=1.0, metric='euclidean'):
-    '''
-    Wrapper for the maximum margin contrastive loss (Hadsell et al. 2006)
-    `tfa.losses.contrastive_loss`
-    Args:
-        z: hidden vector of shape [bsz, n_features].
-        y: ground truth of shape [bsz].
-        metric: one of ('euclidean', 'cosine')
-    '''
-    # compute pair-wise distance matrix
-    if metric == 'euclidean':
-        D = pdist_euclidean(z)
-    elif metric == 'cosine':
-        D = 1 - tf.matmul(z, z, transpose_a=False, transpose_b=True)
-    # convert squareform matrix to vector form
-    d_vec = square_to_vec(D)
-    # make contrastive labels
-    y_contrasts = get_contrast_batch_labels(y)
-    loss = contrastive_loss(y_contrasts, d_vec, margin=margin)
-    # exploding/varnishing gradients on large batch?
-    return tf.reduce_mean(loss)
-# ------
+supConloss = SupConLoss()
 
 @tf.function
 def train_step(images, labels):
 	with tf.GradientTape() as tape:
 		r = encoder_r(images, training=True)
 		z = projector_z(r, training=True)
-		loss = max_margin_contrastive_loss(z, labels)
+		loss = supConloss(z, labels)
 
 	gradients = tape.gradient(loss, 
 		encoder_r.trainable_variables + projector_z.trainable_variables)
@@ -237,37 +120,34 @@ def train_step(images, labels):
 
 	return loss
 
-EPOCHS = 15
+EPOCHS = 20
 LOG_EVERY = 1
 train_loss_results = []
 
-start = time.time()
-for epoch in range(EPOCHS):	
-	epoch_loss_avg = tf.keras.metrics.Mean()
-	
-	for (images, labels) in train_ds:
-		loss = train_step(images, labels)
-		epoch_loss_avg.update_state(loss) 
+print('[training representations]')
+with time_counter():
+    for epoch in range(EPOCHS):	
+        epoch_loss_avg = tf.keras.metrics.Mean()
+        
+        for (images, labels) in train_ds:
+            loss = train_step(images, labels)
+            epoch_loss_avg.update_state(loss) 
 
-	train_loss_results.append(epoch_loss_avg.result())
-	# print({"supervised_contrastive_loss": epoch_loss_avg.result()})
+            # TODO loss が nan になることがある
+            if tf.math.is_nan(loss):
+                print("loss", loss)
+                print("encoder_r output is nan:", tf.reduce_any(tf.math.is_nan(encoder_r(images))))
+                print("projector_z output is nan:", tf.reduce_any(tf.math.is_nan(projector_z(encoder_r(images)))))
+                sys.exit()
 
-	if epoch % LOG_EVERY == 0:
-		print("Epoch: {} Loss: {:.3f}".format(epoch, epoch_loss_avg.result()))
+        train_loss_results.append(epoch_loss_avg.result())
+        # print({"supervised_contrastive_loss": epoch_loss_avg.result()})
 
-end = time.time()
-print({"training_time": end - start})
+        if epoch % LOG_EVERY == 0:
+            print("Epoch: {} Loss: {:.3f}".format(epoch, epoch_loss_avg.result()))
 
-
-# import matplotlib.pyplot as plt
-# plt.plot(train_loss_results)
-# plt.title("Supervised Contrastive Loss")
-# plt.show()
-# # print(train_loss_results)
 supCon_file_path = str(result_path / 'supCon.csv')
 pd.DataFrame(train_loss_results, columns=['supervised_contrastive_loss']).to_csv(supCon_file_path)
-
-
 
 
 # =============================================================================
@@ -288,7 +168,6 @@ def supervised_model():
 	return supervised_model
 
 model = supervised_model()
-
 
 # Loss
 loss = tfk.losses.SparseCategoricalCrossentropy()
